Wrap OpenMP directives for Windows compatibility

+++ src/hardware/cpu/nn/cpu_losses.cpp
@@ -31,7 +31,7 @@ float cpu_categorical_cross_entropy(Tensor* y_true, Tensor* y_pred){
     float eps = 10e-8;
 
     #pragma omp parallel for reduction(+:sum)
-    for (unsigned int bi = 0; bi<y_true->shape[0]; bi++) {  // Batches
+    for (long long int bi = 0; bi<y_true->shape[0]; bi++) {  // Batches
         unsigned int step_i = bi * y_true->stride[0];
 
         // Compute cross-entropy
@@ -51,7 +51,7 @@ void cpu_d_categorical_cross_entropy(Tensor* y_true, Tensor* y_pred, Tensor* del
     float eps = 10e-8;
 
     #pragma omp parallel for
-    for (unsigned int i = 0; i<y_true->size; i++) {
+    for (long long int i = 0; i<y_true->size; i++) {
         delta->ptr[i] = -y_true->ptr[i] * (1.0f/ (y_pred->ptr[i]+eps) );
     }
 }
@@ -61,7 +61,7 @@ float cpu_binary_cross_entropy(Tensor* y_true, Tensor* y_pred){
     float eps = 10e-8;
 
     #pragma omp parallel for reduction(+:sum)
-    for (unsigned int i = 0; i<y_true->size; i++) {
+    for (long long int i = 0; i < y_true->size; i++) {
         sum += y_true->ptr[i] * ::logf(y_pred->ptr[i]+eps) + (1.0-y_true->ptr[i]) * ::logf(1.0f-y_pred->ptr[i]+eps);
     }
 
@@ -74,7 +74,7 @@ void cpu_d_binary_cross_entropy(Tensor* y_true, Tensor* y_pred, Tensor* delta){
     float eps = 10e-8;
 
     #pragma omp parallel for
-    for (unsigned int i = 0; i<y_true->size; i++) {
+    for (long long int i = 0; i<y_true->size; i++) {
         delta->ptr[i] = -( y_true->ptr[i] * 1.0f/(y_pred->ptr[i]+eps) + (1.0-y_true->ptr[i]) * 1.0f/(1.0f-y_pred->ptr[i]+eps) * -1.0f );
     }
 }
